primecolada-backend/requirements.txt
Flask
Flask-Cors
google-cloud-firestore
requests
Flask-Swagger-UI
marshmallow
python-dotenv
Flask-JWT-Extended
gunicorn
